+++ src/Symfony/Component/Form/Extension/Core/DataTransformer/DateTimeToRfc3339Transformer.php
@@ -11,84 +11,18 @@
 
 namespace Symfony\Component\Form\Extension\Core\DataTransformer;
 
-use Symfony\Component\Form\Exception\TransformationFailedException;
-
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
+ *
+ * @deprecated The Symfony\Component\Form\Extension\Core\DataTransformer\DateTimeToRfc3339Transformer class is deprecated since version 4.1 and will be removed in 5.0. Use the Symfony\Component\Form\Extension\Core\DataTransformer\DateTimeImmutableToRfc3339Transformer class instead.
  */
 class DateTimeToRfc3339Transformer extends BaseDateTimeTransformer
 {
-    /**
-     * Transforms a normalized date into a localized date.
-     *
-     * @param \DateTimeInterface $dateTime A DateTimeInterface object
-     *
-     * @return string The formatted date
-     *
-     * @throws TransformationFailedException If the given value is not a \DateTimeInterface
-     */
-    public function transform($dateTime)
+    use DateTimeImmutableTransformerDecoratorTrait;
+
+    public function __construct($inputTimezone = null, $outputTimezone = null)
     {
-        if (null === $dateTime) {
-            return '';
-        }
-
-        if (!$dateTime instanceof \DateTimeInterface) {
-            throw new TransformationFailedException('Expected a \DateTimeInterface.');
-        }
-
-        if ($this->inputTimezone !== $this->outputTimezone) {
-            if (!$dateTime instanceof \DateTimeImmutable) {
-                $dateTime = clone $dateTime;
-            }
-
-            $dateTime = $dateTime->setTimezone(new \DateTimeZone($this->outputTimezone));
-        }
-
-        return preg_replace('/\+00:00$/', 'Z', $dateTime->format('c'));
-    }
-
-    /**
-     * Transforms a formatted string following RFC 3339 into a normalized date.
-     *
-     * @param string $rfc3339 Formatted string
-     *
-     * @return \DateTime Normalized date
-     *
-     * @throws TransformationFailedException If the given value is not a string,
-     *                                       if the value could not be transformed
-     */
-    public function reverseTransform($rfc3339)
-    {
-        if (!is_string($rfc3339)) {
-            throw new TransformationFailedException('Expected a string.');
-        }
-
-        if ('' === $rfc3339) {
-            return;
-        }
-
-        try {
-            $dateTime = new \DateTime($rfc3339);
-        } catch (\Exception $e) {
-            throw new TransformationFailedException($e->getMessage(), $e->getCode(), $e);
-        }
-
-        if ($this->inputTimezone !== $dateTime->getTimezone()->getName()) {
-            $dateTime->setTimezone(new \DateTimeZone($this->inputTimezone));
-        }
-
-        if (preg_match('/(\d{4})-(\d{2})-(\d{2})/', $rfc3339, $matches)) {
-            if (!checkdate($matches[2], $matches[3], $matches[1])) {
-                throw new TransformationFailedException(sprintf(
-                    'The date "%s-%s-%s" is not a valid date.',
-                    $matches[1],
-                    $matches[2],
-                    $matches[3]
-                ));
-            }
-        }
-
-        return $dateTime;
+        parent::__construct($inputTimezone, $outputTimezone);
+        $this->decorated = new DateTimeImmutableToRfc3339Transformer($inputTimezone, $outputTimezone);
     }
 }
